requirements.txt
adal==1.2.2
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
azure-cognitiveservices-language-luis==0.5.0
azure-common==1.1.25
botbuilder-core==4.8.0
botbuilder-schema==4.8.0
botframework-connector==4.8.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
idna==2.9
importlib-metadata==1.6.0
isodate==0.6.0
jsonpickle==1.4
msal==1.2.0
msrest==0.6.13
msrestazure==0.6.3
multidict==4.7.5
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
yarl==1.4.2
zipp==3.1.0
